fix: operand is type `sbyte`

+++ lc-hax/Scripts/Patches/InfiniteDepositPatch.cs
@@ -6,7 +6,7 @@ using HarmonyLib;
 class InfiniteDepositPatch {
     static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions) {
         foreach (CodeInstruction instruction in instructions) {
-            yield return instruction.opcode == OpCodes.Ldc_I4_S && instruction.operand.Equals(12)
+            yield return instruction.opcode == OpCodes.Ldc_I4_S && instruction.operand.Equals((sbyte)12)
                 ? new CodeInstruction(OpCodes.Ldc_I4, int.MaxValue)
                 : instruction;
         }
